Created enemy
Created the enemy, animated him, and set up some collisions to keep him from walking off the ledge. Unfortunately, he seems to be colliding with the ladder (and getting stuck in the middle) at this point , but I've also got somewhat weird behavior with the player and the ladder so I'm not surprised. Hopefully I can get some help online.